==> 1_api_use.ipynb <==
import requests
from bs4 import BeautifulSoup as BS

import pandas as pd

from tqdm.auto import tqdm

api_key = ''
headers = {'Authorization' : api_key}

# nick_params = {'nickname' : '아케인케인'}
# user_url = 'https://api.nexon.co.kr/fifaonline4/v1.0/users?'

# user_info = requests.get(user_url, params=nick_params, headers=headers).json()
# user_df = pd.DataFrame(user_info, index=[0])

# user_df

pcode_url = 'https://static.api.nexon.co.kr/fifaonline4/latest/spid.json'

pcode_info = requests.get(pcode_url, headers=headers).json()
player_df = pd.DataFrame(pcode_info, index=[i for i in range(len(pcode_info))])

pcode_info

player_df

player_df.info()

# df 행 삭제
# pcode_df.drop(0, inplace=True)
# pcode_df.reset_index(inplace=True)
# pcode_df.drop('index', axis=1, inplace=True)

# 이미지 한번에 다운로드
# for x in tqdm(player_df['id'].tolist()):
#     stand_img_url = f'https://fo4.dn.nexoncdn.co.kr/live/externalAssets/common/players/p{x}.png'
#     action_img_url = f'https://fo4.dn.nexoncdn.co.kr/live/externalAssets/common/playersAction/p{x}.png'
#     stand_img_path = f'C:/Users/rladb/workspace/data/fifafow/standimg/{x}.png'
#     action_img_path = f'C:/Users/rladb/workspace/data/fifafow/actionimg/{x}.png'
#     stand_img_file = requests.get(stand_img_url)
#     action_img_file = requests.get(action_img_url)
#     with open(stand_img_path, 'wb') as sphoto:
#         sphoto.write(stand_img_file.content)
#     with open(stand_img_path, 'wb') as aphoto:
#         aphoto.write(stand_img_file.content)

for x in player_df:
    player_df['season_id'] = player_df['id'] / 1000000

player_df['id'] = player_df['id'].astype(int)
player_df['season_id'] = player_df['season_id'].astype(int)

for x in player_df:
    player_df['class'] = player_df['season_id'].map({101:"ICON", 201:"NHD", 202:"TKI", 206:"TB", 207:"TT", \
        210:"GR", 211:"19TOTY", 212:"18TOTY", 213:"MCICON", 214:"TC", 215:"19TOTS", 216:"HOT", 217:"COC", 218:"OTW", \
        219:"19NG", 220:"20TOTY", 221:"19UCL", 222:"20TOTN", 225:"TKL", 230:"20TOTS", 231:"VTR", 233:"MOG", 234:"LH", \
        236:"LA", 237:"MC", 238:"20NG", 239:"20KLB", 240:"21TOTY", 241:"21TOTN", 242:"20UCL", 246:"UP", 247:"2012KH", \
        249:"NTG", 250:"21TOTS", 251:"EBS", 252:"CAP", 253:"BOE21", 254:"CFA", 256:"BTB", 257:"21KLB", 258:"22TOTY", \
        259:"22TOTN", 260:"21UCL", 261:"22HR", 262:"21NG", 264:"FA", 265:"LOL", 267:"22TOTS", 268:"LN", 270:"SPL", \
        272:"BWC", 273:"WC22", 274:"RMCF", 276:"22NG", 277:"22KLB", 278:"23TOTY", 279:"23TOTN", 280:"22UCL", 281:"23HR", \
        283:"HG", 284:"RTN", \
        293:"22KFA", 294:"21KFA", 295:"19KFA", 297:"MCFC", 298:"18KFA", 300:"22LIVE", 317:"17LIVE", 318:"18LIVE", \
        319:"19LIVE", 320:"20LIVE", 321:"21LIVE", 500:"18PLA", 501:"18PLS", 502:"19PLA", 503:"19PLS", 504:"20KL", \
        506:"20PL", 507:"21KL", 508:"21PL", 510:"22KL", 511:"22PL"})

player_df.drop('season_id', axis=1, inplace=True)

player_df

df_10001 = player_df.iloc[:10001]
df_10001

df_10001_20001 = player_df.iloc[10001:20001]
df_10001_20001

df_20001_30001 = player_df.iloc[20001:30001]
df_20001_30001

df_30001_40001 = player_df.iloc[30001:40001]
df_30001_40001

df_40001_50001 = player_df.iloc[40001:50001]
df_40001_50001

df_61990 = player_df.iloc[50001:]
df_61990

player_df.to_csv('C:/Users/rladb/workspace/data/fifafow/playerdf.csv', header=True, index=False, encoding='utf-8-sig')

df_10001.to_csv('C:/Users/rladb/workspace/data/fifafow/df_10001.csv', header=True, index=False, encoding='utf-8-sig')

df_10001_20001.to_csv('C:/Users/rladb/workspace/data/fifafow/df_10001_20001.csv', header=True, index=False, encoding='utf-8-sig')

df_20001_30001.to_csv('C:/Users/rladb/workspace/data/fifafow/df_20001_30001.csv', header=True, index=False, encoding='utf-8-sig')

df_30001_40001.to_csv('C:/Users/rladb/workspace/data/fifafow/df_30001_40001.csv', header=True, index=False, encoding='utf-8-sig')

df_40001_50001.to_csv('C:/Users/rladb/workspace/data/fifafow/df_40001_50001.csv', header=True, index=False, encoding='utf-8-sig')

df_61990.to_csv('C:/Users/rladb/workspace/data/fifafow/df_61990.csv', header=True, index=False, encoding='utf-8-sig')
