==> 2_stats_crawler.ipynb <==
import requests
from bs4 import BeautifulSoup as BS
from user_agent import generate_user_agent, generate_navigator

import pandas as pd

from tqdm.auto import tqdm
import re

crawled_df = pd.read_csv('C:/Users/rladb/workspace/data/fifafow/crawleddf.csv')

crawled_df

player_df = pd.read_csv('C:/Users/rladb/workspace/data/fifafow/playerdf.csv')
df_10001 = pd.read_csv('C:/Users/rladb/workspace/data/fifafow/df_10001.csv')
df_10001_20001 = pd.read_csv('C:/Users/rladb/workspace/data/fifafow/df_10001_20001.csv')
df_20001_30001 = pd.read_csv('C:/Users/rladb/workspace/data/fifafow/df_20001_30001.csv')
df_30001_40001 = pd.read_csv('C:/Users/rladb/workspace/data/fifafow/df_30001_40001.csv')
df_40001_50001 = pd.read_csv('C:/Users/rladb/workspace/data/fifafow/df_40001_50001.csv')
df_61990 = pd.read_csv('C:/Users/rladb/workspace/data/fifafow/df_61990.csv')

player_df

def dictsetter():
    temp = {'pay': [], 'ovr': [], 'position': [], 'age': [], 'height': [], 'weight': [], 'bodytype': [], \
            'skill': [], 'leftfoot': [], 'rightfoot': [], 'reputation': [], 'nation': [], 'o_pac': [], \
            'o_sho': [], 'o_pas': [], 'o_dri': [], 'o_def': [], 'o_phy': [], 'sprint': [], 'accel': [], \
            'finish': [], 'shotpower': [], 'longshot': [], 'positioning': [], 'volley': [], 'penalty': [], \
            'shortpass': [], 'vision': [], 'cross': [], 'longpass': [], 'freekick': [], 'curve': [], 'dribble': [], \
            'control': [], 'agility': [], 'balance': [], 'reaction': [], 'marking': [], 'tackle': [], \
            'intercept': [], 'header': [], 'sliding': [], 'strength': [], 'stamina': [], 'aggression': [], \
            'jump': [], 'composure': [], 'gk_diving': [], 'gk_handle': [], 'gk_kick': [], 'gk_reflex': [], \
            'gk_positioning': [], }
    return temp

def crawler (df):
    result = dictsetter()
    for x in tqdm(df['id'].tolist()):
        url = f'https://fifaonline4.nexon.com/DataCenter/PlayerInfo?spid={x}&n1Strong=0'
        ranheaders = {'User-Agent': generate_user_agent(os='win', device_type='desktop')}
        html = requests.get(url, headers=ranheaders).text
        soup = BS(html, 'html.parser')
        body = soup.find('body')

        info_wrap = body.find('div', class_='info_wrap')
        middle = body.find('div', class_='content_middle').find('ul')
        middle_values = middle.findAll('div', class_=re.compile('value +'))
        bottom = body.find('div', class_='content_bottom')
        bottom_wrap = bottom.findAll('ul', class_=re.compile('data_wrap+'))

        result['pay'].append(info_wrap.find('div', class_='pay_side').text.strip())
        result['ovr'].append(body.find('div', class_='ovr value').text.strip())
        result['position'].append(body.find('div', class_='position').text.strip())
        result['age'].append(info_wrap.find('span', class_='etc birth').text.strip()[-4:-2])
        result['height'].append(info_wrap.find('span', class_='etc height').text.strip()[:3])
        result['weight'].append(info_wrap.find('span', class_='etc weight').text.strip()[:2])
        result['bodytype'].append(info_wrap.find('span', class_='etc physical').text.strip())
        result['skill'].append(len(info_wrap.find('span', class_='etc skill').find('span').text.strip()))
        result['leftfoot'].append(info_wrap.find('span', class_='etc foot').text.strip()[1])
        result['rightfoot'].append(info_wrap.find('span', class_='etc foot').text.strip()[-1])
        result['reputation'].append(info_wrap.find('span', class_='etc season').text.strip())
        result['nation'].append(info_wrap.find('div', class_='etc nation').find('span', class_='txt').text.strip())
        result['o_pac'].append(middle_values[0].text.strip())
        result['o_sho'].append(middle_values[1].text.strip())
        result['o_pas'].append(middle_values[2].text.strip())
        result['o_dri'].append(middle_values[3].text.strip())
        result['o_def'].append(middle_values[4].text.strip())
        result['o_phy'].append(middle_values[5].text.strip())
        result['sprint'].append(bottom_wrap[0].findAll('div', class_=re.compile('value+'))[0].text.strip())
        result['accel'].append(bottom_wrap[0].findAll('div', class_=re.compile('value+'))[1].text.strip())
        result['finish'].append(bottom_wrap[0].findAll('div', class_=re.compile('value+'))[2].text.strip())
        result['shotpower'].append(bottom_wrap[0].findAll('div', class_=re.compile('value+'))[3].text.strip())
        result['longshot'].append(bottom_wrap[0].findAll('div', class_=re.compile('value+'))[4].text.strip())
        result['positioning'].append(bottom_wrap[0].findAll('div', class_=re.compile('value+'))[5].text.strip())
        result['volley'].append(bottom_wrap[0].findAll('div', class_=re.compile('value+'))[6].text.strip())
        result['penalty'].append(bottom_wrap[1].findAll('div', class_=re.compile('value+'))[0].text.strip())
        result['shortpass'].append(bottom_wrap[1].findAll('div', class_=re.compile('value+'))[1].text.strip())
        result['vision'].append(bottom_wrap[1].findAll('div', class_=re.compile('value+'))[2].text.strip())
        result['cross'].append(bottom_wrap[1].findAll('div', class_=re.compile('value+'))[3].text.strip())
        result['longpass'].append(bottom_wrap[1].findAll('div', class_=re.compile('value+'))[4].text.strip())
        result['freekick'].append(bottom_wrap[1].findAll('div', class_=re.compile('value+'))[5].text.strip())
        result['curve'].append(bottom_wrap[1].findAll('div', class_=re.compile('value+'))[6].text.strip())
        result['dribble'].append(bottom_wrap[2].findAll('div', class_=re.compile('value+'))[0].text.strip())
        result['control'].append(bottom_wrap[2].findAll('div', class_=re.compile('value+'))[1].text.strip())
        result['agility'].append(bottom_wrap[2].findAll('div', class_=re.compile('value+'))[2].text.strip())
        result['balance'].append(bottom_wrap[2].findAll('div', class_=re.compile('value+'))[3].text.strip())
        result['reaction'].append(bottom_wrap[2].findAll('div', class_=re.compile('value+'))[4].text.strip())
        result['marking'].append(bottom_wrap[2].findAll('div', class_=re.compile('value+'))[5].text.strip())
        result['tackle'].append(bottom_wrap[2].findAll('div', class_=re.compile('value+'))[6].text.strip())
        result['intercept'].append(bottom_wrap[3].findAll('div', class_=re.compile('value+'))[0].text.strip())
        result['header'].append(bottom_wrap[3].findAll('div', class_=re.compile('value+'))[1].text.strip())
        result['sliding'].append(bottom_wrap[3].findAll('div', class_=re.compile('value+'))[2].text.strip())
        result['strength'].append(bottom_wrap[3].findAll('div', class_=re.compile('value+'))[3].text.strip())
        result['stamina'].append(bottom_wrap[3].findAll('div', class_=re.compile('value+'))[4].text.strip())
        result['aggression'].append(bottom_wrap[3].findAll('div', class_=re.compile('value+'))[5].text.strip())
        result['jump'].append(bottom_wrap[3].findAll('div', class_=re.compile('value+'))[6].text.strip())
        result['composure'].append(bottom_wrap[4].findAll('div', class_=re.compile('value+'))[0].text.strip())
        result['gk_diving'].append(bottom_wrap[4].findAll('div', class_=re.compile('value+'))[1].text.strip())
        result['gk_handle'].append(bottom_wrap[4].findAll('div', class_=re.compile('value+'))[2].text.strip())
        result['gk_kick'].append(bottom_wrap[4].findAll('div', class_=re.compile('value+'))[3].text.strip())
        result['gk_reflex'].append(bottom_wrap[4].findAll('div', class_=re.compile('value+'))[4].text.strip())
        result['gk_positioning'].append(bottom_wrap[4].findAll('div', class_=re.compile('value+'))[5].text.strip())

    temp_df = pd.DataFrame(data = result)
    return temp_df

crawled_df = crawler(player_df)
crawled_df

crawled_df = pd.concat([player_df, crawled_df], axis=1)
crawled_df

temp_nation = []

for i in range(0, 61991):
    x = crawled_df['nation'].iloc[i].replace(', 국가대표', '')
    temp_nation.append(x)

crawled_df['nation'] = temp_nation

crawled_df.to_csv('C:/Users/rladb/workspace/data/fifafow/crawleddf.csv', header=True, index=False, encoding='utf-8-sig')

# info_wrap = body.find('div', class_='info_wrap')
# middle = body.find('div', class_='content_middle').find('ul')
# middle_values = middle.findAll('div', class_=re.compile('value +'))
# bottom = body.find('div', class_='content_bottom')
# bottom_wrap = bottom.findAll('ul', class_=re.compile('data_wrap+'))
#pay
# body.find('div', class_='pay').text.strip()
#ovr
# body.find('div', class_='ovr value').text.strip()
#position
# body.find('div', class_='position').text.strip()
#age
# info_wrap.find('span', class_='etc birth').text.strip()[-4:-2]
#height
# info_wrap.find('span', class_='etc height').text.strip()[:3]
#weight
# info_wrap.find('span', class_='etc weight').text.strip()[:2]
#bodytype
# info_wrap.find('span', class_='etc physical').text.strip()
#skill
# len(info_wrap.find('span', class_='etc skill').find('span').text.strip())
#leftfoot
# info_wrap.find('span', class_='etc foot').text.strip()[1]
#rightfoot
# info_wrap.find('span', class_='etc foot').text.strip()[-1]
#reputation
# info_wrap.find('span', class_='etc season').text.strip()
#nation
# info_wrap.find('div', class_='etc nation').find('span', class_='txt').text.strip()
#ovr 6 stats
# middle.findAll('div', class_=re.compile('value +'))[0].text.strip()
#34 stats (first 7)
# bottom_wrap[0].findAll('div', class_=re.compile('value+'))[0].text.strip()
