==> 3_mugshot_append.ipynb <==
import requests
from bs4 import BeautifulSoup as BS
from user_agent import generate_user_agent, generate_navigator

import pandas as pd

from tqdm.auto import tqdm

crawled_df = pd.read_csv('C:/Users/rladb/workspace/data/fifafow/crawleddf.csv')

df_10001 = pd.read_csv('C:/Users/rladb/workspace/data/fifafow/df_10001.csv')
df_10001_20001 = pd.read_csv('C:/Users/rladb/workspace/data/fifafow/df_10001_20001.csv')
df_20001_30001 = pd.read_csv('C:/Users/rladb/workspace/data/fifafow/df_20001_30001.csv')
df_30001_40001 = pd.read_csv('C:/Users/rladb/workspace/data/fifafow/df_30001_40001.csv')
df_40001_50001 = pd.read_csv('C:/Users/rladb/workspace/data/fifafow/df_40001_50001.csv')
df_61990 = pd.read_csv('C:/Users/rladb/workspace/data/fifafow/df_61990.csv')

crawled_df

def mugshot_appender(df):
    temp_mugshot = []
    for x in tqdm(df['id']):
        url = f'https://fo4.dn.nexoncdn.co.kr/live/externalAssets/common/playersAction/p{x}.png'
        ranheaders = {'User-Agent': generate_user_agent(os='win', device_type='desktop')}
        status = requests.get(url, headers=ranheaders).status_code
        if (status == 200):
            temp_mugshot.append(f'https://fo4.dn.nexoncdn.co.kr/live/externalAssets/common/playersAction/p{x}.png')
        else:
            y = int(str(x)[3:])
            temp_mugshot.append(f'https://fo4.dn.nexoncdn.co.kr/live/externalAssets/common/players/p{y}.png')
    return temp_mugshot

mugshot_10001 = mugshot_appender(df_10001)

mugshot_10001_20001 = mugshot_appender(df_10001_20001)

mugshot_20001_30001 = mugshot_appender(df_20001_30001)

mugshot_30001_40001 = mugshot_appender(df_30001_40001)

mugshot_40001_50001 = mugshot_appender(df_40001_50001)

mugshot_61990 = mugshot_appender(df_61990)

mugshot = mugshot_10001 + mugshot_10001_20001 + mugshot_20001_30001 + mugshot_30001_40001 + mugshot_40001_50001 + mugshot_61990

len(mugshot)

crawled_df['mugshot'] = mugshot

crawled_df.info()

crawled_df.to_csv('C:/Users/rladb/workspace/data/fifafow/crawleddf.csv', header=True, index=False, encoding='utf-8-sig')
